requirements.txt
biopython==1.72
boto==2.49.0
boto3==1.7.56
botocore==1.10.56
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
colorlover==0.2.1
decorator==4.3.0
docutils==0.14
gensim==3.5.0
idna==2.7
ipython-genutils==0.2.0
jmespath==0.9.3
jsonschema==2.6.0
jupyter-core==4.4.0
nbformat==4.4.0
networkx==2.1
numpy==1.14.5
plotly==3.0.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
retrying==1.3.3
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
smart-open==1.6.0
traitlets==4.3.2
urllib3==1.23
